+++ src/Annotations/tests/Fixtures/Annotation/MethodAnnotation.php
@@ -11,10 +11,13 @@ declare(strict_types=1);
 
 namespace Spiral\Tests\Annotations\Fixtures\Annotation;
 
+use Attribute;
+
 /**
  * @Annotation
  */
-class Route
+#[Attribute(Attribute::TARGET_METHOD)]
+class MethodAnnotation
 {
     /** @var string */
     public $path;