+++ src/Models/src/AbstractEntity.php
@@ -199,7 +199,7 @@ abstract class AbstractEntity implements EntityInterface, ValueInterface, \Itera
     /**
      * {@inheritdoc}
      */
-    public function offsetExists($offset)
+    public function offsetExists($offset): bool
     {
         return $this->__isset($offset);
     }
@@ -207,6 +207,7 @@ abstract class AbstractEntity implements EntityInterface, ValueInterface, \Itera
     /**
      * {@inheritdoc}
      */
+    #[\ReturnTypeWillChange]
     public function offsetGet($offset)
     {
         return $this->getField($offset);