+++ src/Pagination/src/Paginator.php
@@ -130,10 +130,7 @@ final class Paginator implements PaginatorInterface, \Countable
         return $paginator;
     }
 
-    /**
-     * @return int
-     */
-    public function count()
+    public function count(): int
     {
         return $this->count;
     }