+++ src/Reactor/src/Aggregator.php
@@ -140,7 +140,7 @@ class Aggregator extends AbstractDeclaration implements
     /**
      * {@inheritdoc}
      */
-    public function offsetExists($offset)
+    public function offsetExists($offset): bool
     {
         return $this->has($offset);
     }
@@ -148,6 +148,7 @@ class Aggregator extends AbstractDeclaration implements
     /**
      * {@inheritdoc}
      */
+    #[\ReturnTypeWillChange]
     public function offsetGet($offset)
     {
         return $this->get($offset);