+++ src/Stempler/src/Lexer/Grammar/Dynamic/DirectiveGrammar.php
@@ -98,7 +98,7 @@ final class DirectiveGrammar implements \IteratorAggregate
      *
      * @return \Generator|\Traversable
      */
-    public function getIterator()
+    public function getIterator(): \Traversable
     {
         if ($this->tokens === []) {
             throw new \LogicException('Directive not parsed');